Rename gaussian mixture and improve doc
