+++ moe/src/lib.rs
@@ -86,7 +86,6 @@ mod clustering;
 mod errors;
 mod expertise_macros;
 mod gaussian_mixture;
-mod multivariate_normal;
 mod parameters;
 mod surrogates;
 
@@ -94,6 +93,5 @@ pub use algorithm::*;
 pub use clustering::*;
 pub use errors::*;
 pub use gaussian_mixture::*;
-pub use multivariate_normal::*;
 pub use parameters::*;
 pub use surrogates::*;
